edits to modal and small changes
