+++ react-app/src/components/LogInModal/index.js
@@ -5,12 +5,14 @@ import LoginForm from './LoginForm';
 function LoginFormModal() {
     const [showModal, setShowModal] = useState(false);
 
+    const onClose = () => setShowModal(false)
+
     return (
         <>
             <button onClick={() => setShowModal(true)}>Sign In</button>
             {showModal && (
-                <Modal onClose={() => setShowModal(false)}>
-                    <LoginForm />
+                <Modal onClose={onClose}>
+                    <LoginForm onClose={onClose}/>
                 </Modal>
             )}
         </>
